yoloImageDetector/requirements.txt
flask
ultralytics
torch
opencv-python-headless
Pillow
gunicorn